Fix #5 (wrong proportions in maximized view)
The key is to call the extended constructor. Determining the max dimension is the tricky part because some graphs are looking to empty on big screens if the factor is 1. With the introduction of the 'factor' variable it is possible to quickly adjust the scaling.
Remaining problem: if the factor is not 0.5 or 1, markers (especially circles) are looking a bit weird. Not really distorted, more like a few pixels are cut off at some sides.